api/image_preprocessing/requirements.txt
azure-functions
IMAGE